Refactor acuant a/b testing and csp into concern
