+++ app/controllers/concerns/idv/hybrid_mobile/hybrid_mobile_concern.rb
@@ -3,32 +3,10 @@ module Idv
     module HybridMobileConcern
       extend ActiveSupport::Concern
 
+      include AcuantConcern
+
       included do
         before_action :render_404_if_hybrid_mobile_controllers_disabled
-      end
-
-      def acuant_sdk_ab_test_analytics_args
-        {
-          acuant_sdk_upgrade_ab_test_bucket:
-            AbTests::ACUANT_SDK.bucket(document_capture_session_uuid),
-        }
-      end
-
-      def acuant_sdk_upgrade_a_b_testing_variables
-        bucket = AbTests::ACUANT_SDK.bucket(document_capture_session_uuid)
-        testing_enabled = IdentityConfig.store.idv_acuant_sdk_upgrade_a_b_testing_enabled
-        use_alternate_sdk = (bucket == :use_alternate_sdk)
-        if use_alternate_sdk
-          acuant_version = IdentityConfig.store.idv_acuant_sdk_version_alternate
-        else
-          acuant_version = IdentityConfig.store.idv_acuant_sdk_version_default
-        end
-        {
-          acuant_sdk_upgrade_a_b_testing_enabled:
-              testing_enabled,
-          use_alternate_sdk: use_alternate_sdk,
-          acuant_version: acuant_version,
-        }
       end
 
       def check_valid_document_capture_session
