+++ app/controllers/concerns/idv/step_utilities_concern.rb
@@ -1,20 +1,7 @@
 module Idv
   module StepUtilitiesConcern
     extend ActiveSupport::Concern
-
-    # Copied from capture_doc_flow.rb
-    # and from doc_auth_flow.rb
-    def acuant_sdk_ab_test_analytics_args
-      capture_session_uuid = flow_session[:document_capture_session_uuid]
-      if capture_session_uuid
-        {
-          acuant_sdk_upgrade_ab_test_bucket:
-            AbTests::ACUANT_SDK.bucket(capture_session_uuid),
-        }
-      else
-        {}
-      end
-    end
+    include AcuantConcern
 
     def irs_reproofing?
       effective_user&.reproof_for_irs?(
@@ -22,14 +9,15 @@ module Idv
       ).present?
     end
 
-    def document_capture_session
-      @document_capture_session ||= DocumentCaptureSession.find_by(
-        uuid: flow_session[document_capture_session_uuid_key],
-      )
+    def document_capture_session_uuid
+      flow_session[:document_capture_session_uuid]
     end
 
-    def document_capture_session_uuid_key
-      :document_capture_session_uuid
+    def document_capture_session
+      return @document_capture_session if defined?(@document_capture_session)
+      @document_capture_session = DocumentCaptureSession.find_by(
+        uuid: document_capture_session_uuid,
+      )
     end
   end
 end
