+++ app/controllers/idv/document_capture_controller.rb
@@ -1,5 +1,6 @@
 module Idv
   class DocumentCaptureController < ApplicationController
+    include AcuantConcern
     include IdvSession
     include IdvStepConcern
     include StepIndicatorConcern
@@ -75,16 +76,6 @@ module Idv
       redirect_to idv_ssn_url
     end
 
-    # This is copied from DocumentCaptureConcern, without the step check
-    def override_csp_to_allow_acuant
-      policy = current_content_security_policy
-      policy.connect_src(*policy.connect_src, 'us.acas.acuant.net')
-      policy.script_src(*policy.script_src, :unsafe_eval)
-      policy.style_src(*policy.style_src, :unsafe_inline)
-      policy.img_src(*policy.img_src, 'blob:')
-      request.content_security_policy = policy
-    end
-
     def analytics_arguments
       {
         flow_path: flow_path,
@@ -92,23 +83,6 @@ module Idv
         analytics_id: 'Doc Auth',
         irs_reproofing: irs_reproofing?,
       }.merge(**acuant_sdk_ab_test_analytics_args)
-    end
-
-    def acuant_sdk_upgrade_a_b_testing_variables
-      bucket = AbTests::ACUANT_SDK.bucket(flow_session[:document_capture_session_uuid])
-      testing_enabled = IdentityConfig.store.idv_acuant_sdk_upgrade_a_b_testing_enabled
-      use_alternate_sdk = (bucket == :use_alternate_sdk)
-      if use_alternate_sdk
-        acuant_version = IdentityConfig.store.idv_acuant_sdk_version_alternate
-      else
-        acuant_version = IdentityConfig.store.idv_acuant_sdk_version_default
-      end
-      {
-        acuant_sdk_upgrade_a_b_testing_enabled:
-            testing_enabled,
-        use_alternate_sdk: use_alternate_sdk,
-        acuant_version: acuant_version,
-      }
     end
 
     def in_person_cta_variant_testing_variables
